+++ ArmaForces.ArmaServerManager.Discord/Features/Server/IServerManagerClient.cs
@@ -1,0 +1,12 @@
+﻿using ArmaForces.ArmaServerManager.Discord.Features.Server.DTOs;
+using CSharpFunctionalExtensions;
+
+namespace ArmaForces.ArmaServerManager.Discord.Features.Server
+{
+    public interface IServerManagerClient
+    {
+        Result<ServerStatus> GetServerStatus();
+
+        Result RequestStartServer(ServerStartRequest serverStartRequest);
+    }
+}
